scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Just a uv project to run the scripts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygithub>=2.5.0",
    "semver>=3.0.2",
]
